Changes to codemirror source
Was able to pull the source from the wayback machine. Originally I
searched for the actual github repo, but the files within it weren't
archived. Then I realized that the github pages site was probably
archived, and the source files within it.

Luckily, it was, so I recovered the files from there.